feat(services): "issueJWTokens" service moved from controller file to separate one
